[REFACTOR] base services
