--- src/main/java/com/henry/bookrecommendationsystem/controller/BaseController.java
@@ -1,7 +1,0 @@
-package com.henry.bookrecommendationsystem.controller;
-
-import com.henry.bookrecommendationsystem.service.BaseService;
-
-public interface BaseController<Service extends BaseService> {
-    Service getService();
-}
