+++ src/main/java/com/henry/bookrecommendationsystem/controller/base/BaseController.java
@@ -1,0 +1,7 @@
+package com.henry.bookrecommendationsystem.controller.base;
+
+import com.henry.bookrecommendationsystem.service.base.BaseService;
+
+public interface BaseController<Service extends BaseService> {
+    Service getService();
+}
