+++ src/main/java/com/henry/bookrecommendationsystem/transformer/mapper/base/BaseMapper.java
@@ -1,7 +1,7 @@
-package com.henry.bookrecommendationsystem.transformer.mapper;
+package com.henry.bookrecommendationsystem.transformer.mapper.base;
 
 import com.henry.bookrecommendationsystem.dto.base.BaseDto;
-import com.henry.bookrecommendationsystem.entity.BaseEntity;
+import com.henry.bookrecommendationsystem.entity.base.BaseEntity;
 import org.mapstruct.MappingTarget;
 
 public interface BaseMapper<Entity extends BaseEntity, Dto extends BaseDto> {
